@@ -1,6 +1,8 @@
 {
   "name": "metamodels/attribute_alias",
   "description": "MetaModels attribute for alias",
+  "license": "LGPL-3.0-or-later",
+  "type": "contao-bundle",
   "keywords": [
     "contao",
     "metamodels",
@@ -8,9 +10,6 @@
     "alias",
     "nontranslatable"
   ],
-  "type": "contao-bundle",
-  "homepage": "https://now.metamodel.me/",
-  "license": "LGPL-3.0-or-later",
   "authors": [
     {
       "name": "Christian Schiffler",
@@ -24,27 +23,28 @@
       "role": "Developer"
     }
   ],
+  "homepage": "https://now.metamodel.me/",
   "support": {
     "email": "mail@metamodel.me",
     "issues": "https://github.com/MetaModels/attribute_alias/issues",
     "wiki": "https://de.contaowiki.org/MetaModels",
-    "irc": "irc://irc.freenode.org/contao.mm",
     "source": "https://github.com/MetaModels/attribute_alias"
   },
   "require": {
-    "php": "^7.4",
-    "contao-community-alliance/dc-general": "^2.2.0",
-    "contao/core-bundle": "^4.9.0, <4.13.0",
-    "doctrine/dbal": "^2.11 || ^3.0",
-    "menatwork/contao-multicolumnwizard-bundle": "^3.4",
-    "metamodels/core": "^2.1",
-    "symfony/dependency-injection": "4.4.*",
-    "symfony/event-dispatcher": "4.4.*",
-    "symfony/http-kernel": "~4.4.13"
+    "php": "^8.1",
+    "contao-community-alliance/dc-general": "^2.3.29",
+    "contao/core-bundle": "^4.13.0 <5.0",
+    "doctrine/dbal": "^3.3",
+    "menatwork/contao-multicolumnwizard-bundle": "^3.6",
+    "metamodels/core": "^2.3",
+    "symfony/dependency-injection": "^5.4",
+    "symfony/event-dispatcher": "^5.4",
+    "symfony/http-kernel": "^5.4",
+    "symfony/translation-contracts": "^2.5.2"
   },
   "require-dev": {
     "contao/manager-plugin": "^2.1",
-    "phpcq/all-tasks": "^1.2"
+    "phpcq/runner-bootstrap": "^1.0@dev"
   },
   "autoload": {
     "psr-4": {
@@ -59,17 +59,18 @@
       "MetaModels\\AttributeAliasBundle\\Test\\": "tests/Test"
     }
   },
-  "extra": {
-    "contao-manager-plugin": "MetaModels\\AttributeAliasBundle\\ContaoManager\\Plugin",
-    "branch-alias": {
-      "dev-hotfix/2.2.1": "2.2.1-dev"
-    }
-  },
   "config": {
     "allow-plugins": {
       "contao-components/installer": false,
-      "contao/manager-plugin": false
+      "contao/manager-plugin": false,
+      "php-http/discovery": true
     },
     "sort-packages": true
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-feature/2.3.0": "2.3.x-dev"
+    },
+    "contao-manager-plugin": "MetaModels\\AttributeAliasBundle\\ContaoManager\\Plugin"
   }
 }
