@@ -19,6 +19,8 @@
 
 namespace MetaModels\AttributeAliasBundle;
 
+use MetaModels\AttributeAliasBundle\DependencyInjection\CompilerPass\RegisterAttributeInFilterPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -26,5 +28,9 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 class MetaModelsAttributeAliasBundle extends Bundle
 {
-
+    public function build(ContainerBuilder $container): void
+    {
+        parent::build($container);
+        $container->addCompilerPass(new RegisterAttributeInFilterPass());
+    }
 }
