@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_alias.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,7 +18,7 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
- * @copyright  2012-2022 The MetaModels team.
+ * @copyright  2012-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_alias/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -36,18 +36,19 @@ $GLOBALS['TL_DCA']['tl_metamodel_attribute']['metapalettes']['alias extends _sim
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['force_alias'] = [
-    'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['force_alias'],
-    'exclude'   => true,
-    'inputType' => 'checkbox',
-    'default'   => '1',
-    'sql'       => 'char(1) NOT NULL default \'\'',
-    'eval'      => [
+    'label'       => 'force_alias.label',
+    'description' => 'force_alias.description',
+    'exclude'     => true,
+    'inputType'   => 'checkbox',
+    'sql'         => 'char(1) NOT NULL default \'\'',
+    'eval'        => [
         'tl_class' => 'clr w50 cbx m12'
     ],
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['validAliasCharacters'] = [
-    'label'            => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['validAliasCharacters'],
+    'label'            => 'validAliasCharacters.label',
+    'description'      => 'validAliasCharacters.description',
     'exclude'          => true,
     'inputType'        => 'select',
     'options_callback' => static function () {
@@ -64,11 +65,12 @@ $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['validAliasCharacters'] =
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['slugLocale'] = [
-    'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['slugLocale'],
-    'exclude'   => true,
-    'inputType' => 'text',
-    'sql'       => 'varchar(5) default NULL',
-    'eval'      => [
+    'label'       => 'slugLocale.label',
+    'description' => 'slugLocale.description',
+    'exclude'     => true,
+    'inputType'   => 'text',
+    'sql'         => 'varchar(5) default NULL',
+    'eval'        => [
         'rgxp'      => 'language',
         'maxlength' => 5,
         'nospace'   => true,
@@ -77,46 +79,52 @@ $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['slugLocale'] = [
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['noIntegerPrefix'] = [
-    'exclude'   => true,
-    'inputType' => 'checkbox',
-    'default'   => '1',
-    'sql'       => 'char(1) NOT NULL default \'\'',
-    'eval'      => [
+    'label'       => 'noIntegerPrefix.label',
+    'description' => 'noIntegerPrefix.description',
+    'exclude'     => true,
+    'inputType'   => 'checkbox',
+    'default'     => '1',
+    'sql'         => 'char(1) NOT NULL default \'\'',
+    'eval'        => [
         'tl_class' => 'clr w50 cbx m12'
     ],
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['alias_prefix'] = [
-    'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_prefix'],
-    'exclude'   => true,
-    'inputType' => 'text',
-    'sql'       => 'varchar(255) NOT NULL default \'\'',
-    'eval'      => [
+    'label'       => 'alias_prefix.label',
+    'description' => 'alias_prefix.description',
+    'exclude'     => true,
+    'inputType'   => 'text',
+    'sql'         => 'varchar(255) NOT NULL default \'\'',
+    'eval'        => [
         'rgxp'     => 'alpha',
         'tl_class' => 'clr w50'
     ],
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['alias_postfix'] = [
-    'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_postfix'],
-    'exclude'   => true,
-    'inputType' => 'text',
-    'sql'       => 'varchar(255) NOT NULL default \'\'',
-    'eval'      => [
+    'label'       => 'alias_postfix.label',
+    'description' => 'alias_postfix.description',
+    'exclude'     => true,
+    'inputType'   => 'text',
+    'sql'         => 'varchar(255) NOT NULL default \'\'',
+    'eval'        => [
         'tl_class' => 'w50'
     ],
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_attribute']['fields']['alias_fields'] = [
-    'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['alias_fields'],
-    'exclude'   => true,
-    'inputType' => 'multiColumnWizard',
-    'sql'       => 'blob NULL',
-    'eval'      => [
-        'tl_class'     => 'clr w50',
-        'columnFields' => [
+    'label'       => 'alias_fields.label',
+    'description' => 'alias_fields.description',
+    'exclude'     => true,
+    'inputType'   => 'multiColumnWizard',
+    'sql'         => 'blob NULL',
+    'eval'        => [
+        'tl_class'      => 'clr w50',
+        'useTranslator' => true,
+        'columnFields'  => [
             'field_attribute' => [
-                'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['field_attribute'],
+                'label'     => 'field_attribute.label',
                 'exclude'   => true,
                 'inputType' => 'select',
                 'reference' => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['select_values'],
